@@ -1,10 +1,9 @@
 module github.com/hotmall/reverseproxy
 
-go 1.16
+go 1.19
 
 require (
-	github.com/emicklei/go-restful v2.16.0+incompatible
-	github.com/json-iterator/go v1.1.12 // indirect
+	github.com/emicklei/go-restful/v3 v3.11.1
 	github.com/valyala/bytebufferpool v1.0.0
 	gopkg.in/yaml.v3 v3.0.1
 )
